ディレクトリ構成をエンドポイントに寄せた
