+++ backend/flask/api/attributes/delete.py
