+++ backend/flask/api/categories/delete.py
