+++ backend/flask/api/formats/delete.py
