+++ backend/flask/api/groups/delete.py
