+++ backend/flask/api/types/delete.py
